src/requirements.txt
bidict==0.21.2
click==7.1.2
Flask==1.1.2
Flask-SocketIO==5.0.1
gevent==20.12.1
gevent-websocket==0.10.1
greenlet==0.4.17
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-engineio==4.0.0
python-socketio==5.0.4
RPi.GPIO==0.7.0
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
